Refactor: share EnqueueHostTaskData/NativeHostTask between handler and commands
Move EnqueueHostTaskData and NativeHostTask from the anonymous namespace in
commands.cpp into host_task.hpp so both the native recording path in handler.cpp
and the scheduler path in commands.cpp use the same type and callback rather
than duplicating the pattern.
